requirements.txt
dlib==19.24.9
cmake
opencv-python==4.8.1.78
face_recognition==1.3.0
imutils==0.5.4
scikit-learn==1.3.0
numpy==1.26.3
Pillow==10.0.1
tinydb==4.8.0
